Reduce header pollution of EvaluationSession.h
Quite the rats nest of secondary includes.

Broke out the `using`s from function.h and module.h into their own
header which can be used in the multiple places it's needed.

Like, why did export_stl depend on EvaluationSession.h somehow
indirectly including Feature.h!? And how even...
