+++ src/core/BuiltinContext.h
@@ -2,6 +2,7 @@
 
 #include <string>
 
+#include "core/callables.h"
 #include "core/Context.h"
 
 class Location;
