+++ src/core/Builtins.h
@@ -3,7 +3,6 @@
 #include <string>
 #include <unordered_map>
 #include <vector>
-#include "core/module.h"
 #include "core/Assignment.h"
 
 class AbstractModule;
