+++ src/core/Context.h
@@ -6,9 +6,8 @@
 #include <string>
 #include <vector>
 
+#include "core/callables.h"
 #include "core/ContextFrame.h"
-#include "core/AST.h"
-#include "core/ContextMemoryManager.h"
 
 /**
  * Local handle to a all context objects. This is used to maintain the
