+++ src/core/EvaluationSession.cc
@@ -32,6 +32,9 @@
 
 #include "core/AST.h"
 #include "core/ContextFrame.h"
+#include "core/function.h"
+#include "core/module.h"
+#include "core/Value.h"
 #include "utils/printutils.h"
 
 size_t EvaluationSession::push_frame(ContextFrame *frame)
