+++ src/core/Expression.h
@@ -8,9 +8,11 @@
 #include <vector>
 #include <memory>
 #include <boost/logic/tribool.hpp>
+#include <boost/optional.hpp>
+
 #include "core/Assignment.h"
 #include "core/AST.h"
-#include "core/function.h"
+#include "core/callables.h"
 #include "core/Value.h"
 
 template <class T>
