+++ src/core/ModuleInstantiation.cc
@@ -8,6 +8,7 @@
 #include "utils/compiler_specific.h"
 #include "core/Context.h"
 #include "core/Expression.h"
+#include "core/module.h"
 #include "utils/exceptions.h"
 #include "utils/printutils.h"
