+++ src/core/ScopeContext.h
@@ -6,6 +6,7 @@
 
 #include "core/Arguments.h"
 #include "core/AST.h"
+#include "core/callables.h"
 #include "core/Children.h"
 #include "core/Context.h"
 #include "core/SourceFile.h"
