+++ src/core/TransformNode.cc
@@ -29,6 +29,7 @@
 #include "geometry/linalg.h"
 #include "core/Builtins.h"
 #include "core/Children.h"
+#include "core/module.h"
 #include "core/ModuleInstantiation.h"
 #include "core/Parameters.h"
 #include "core/Value.h"
