+++ src/io/export_wrl.cc
@@ -29,6 +29,7 @@
 #include <memory>
 #include <cstddef>
 
+#include "Feature.h"
 #include "geometry/Geometry.h"
 #include "geometry/PolySet.h"
 #include "geometry/PolySetUtils.h"
